+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
+### Fixed
+- SAM presence observation following PC/SC plugin upgrade.
+### Upgraded
+- Keyple components
+  - keyple-plugin-pcsc-java-lib `2.4.2` -> `2.5.1`
 
 ## [2025.03.21]
 ### Upgraded